Trainstation Rent

+++ src/main/java/monopoly/field/Property.java
@@ -3,7 +3,7 @@ package monopoly.field;
 import monopoly.Player;
 
 public abstract class Property extends Field {
-    protected float price;
+    protected double price;
     protected Player owner;
-    public abstract float getRent();
+    public abstract double getRent();
 }
