+++ src/main/java/monopoly/field/PublicInfrastructure.java
@@ -8,7 +8,7 @@ public class PublicInfrastructure extends Property{
 
 
     @Override
-    public float getRent() {
+    public double getRent() {
         return 0;
     }
 }
